requirements.txt
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
cycler==0.11.0
distlib==0.3.6
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
filelock==3.8.2
fonttools==4.33.3
idna==3.4
importlib-metadata==5.2.0
kiwisolver==1.4.2
Markdown==3.4.1
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
platformdirs==2.6.0
psycopg2==2.9.5
pycodestyle==2.8.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.7
urllib3==1.26.14
virtualenv==20.17.1
zipp==3.11.0